+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## [0.1.10] - Unreleased
 ### Added
 * datetime utilities
-* `Dispatcher` class to manage and run remote jobs.
-* `JobRunner` class to run remote jobs.
+* `Dispatcher` class to manage and run remote jobs
+* `JobRunner` class to run remote jobs
+
+### Updated
+* Environment variables `DATABRICKS_SDK_UPSTREAM` and `DATABRICKS_SDK_UPSTREAM_VERSION` to track laktory metrics as a Databricks partner
 
 ## [0.1.9] - 2024-04-17
 ### Added