+++ CHANGELOG.md
@@ -5,9 +5,10 @@
 * datetime utilities
 * `Dispatcher` class to manage and run remote jobs
 * `JobRunner` class to run remote jobs
-
 ### Updated
 * Environment variables `DATABRICKS_SDK_UPSTREAM` and `DATABRICKS_SDK_UPSTREAM_VERSION` to track laktory metrics as a Databricks partner
+### Fixed
+* Permissions resource dependencies on `DbfsFile` and `WorkspaceFile`.
 
 ## [0.1.9] - 2024-04-17
 ### Added